Fixes #672 prevent crash of gof plot when Inf values are introduced
The crash actually came from the auto scaling of the axes.
Since the data included Inf, ggplot was trying to scale to Inf and crashed (the Inf data point only leads to warnings from ggplot).
For tlf, this means there needs to be 2 layers in handling Inf and NAs: one regarding the data points to plot and one preventing scaling by Inf values