Fixes #909 use limits and ticks from boxplots for range plots
`{ggplot2}` crashes when logticks can't be added on log plots (`Error in unit(yticks$start, "cm")`)
The issue was encountered a while ago for boxplots and was fixed by calculating sensible ticks and limits then. Since range plots use the same data, the boxplot ticks and limits can be re-used.